media: mc: Don't care about unsettable flags in MEDIA_IOC_LINK_SETUP
The implementation of MEDIA_IOC_LINK_SETUP currently requires that all
flags that are not configurable are not modified by userspace. This poses
a problem for adding new flags as programs could not work with links that
have unknown flags even when the use of these flags wouldn't affect the
program.

Ignore the non-settable link flags.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>